Birthday Paradox Simulation Solution
The birthday paradox, also known as the birthday problem, states that
in a random group of 23 people, there is about a 50 percent chance that
two people have the same birthday. Is this really true? There are
multiple reasons why this seems like a paradox. One is that when in a
room with 22 other people, if a person compares his or her birthday
with the birthdays of the other people it would make for only 22
comparisons—only 22 chances for people to share the same birthday.